+++ src/main/java/com/miniprojetspring/Service/Implementation/SprintBacklogServiceImpl.java
@@ -7,6 +7,7 @@ import com.miniprojetspring.Repository.SprintBacklogRepository;
 import com.miniprojetspring.Service.ProjectService;
 import com.miniprojetspring.Service.SprintBacklogService;
 import com.miniprojetspring.payload.SprintBacklogPayload;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 
 import java.util.UUID;
@@ -16,14 +17,16 @@ public class SprintBacklogServiceImpl implements SprintBacklogService {
 
     private final SprintBacklogRepository sprintBacklogRepository;
     private final ProjectService projectServiceImpl;
+    private final ProjectSecurityService projectSecurityService;
 
-    public SprintBacklogServiceImpl(SprintBacklogRepository sprintBacklogRepository, ProjectService projectServiceImpl) {
+    public SprintBacklogServiceImpl(SprintBacklogRepository sprintBacklogRepository, ProjectService projectServiceImpl, ProjectSecurityService projectSecurityService) {
         this.sprintBacklogRepository = sprintBacklogRepository;
         this.projectServiceImpl = projectServiceImpl;
+        this.projectSecurityService = projectSecurityService;
     }
 
-    public SprintBacklog createSprintBacklog(String projectId, SprintBacklogPayload payload) {
-        Project project = projectServiceImpl.getProjectById(projectId);
+    public SprintBacklog createSprintBacklog(SprintBacklogPayload payload) {
+        Project project = projectServiceImpl.getProject();
         if (project == null) {
             throw new NotFoundException("Project not found");
         }
@@ -31,16 +34,34 @@ public class SprintBacklogServiceImpl implements SprintBacklogService {
     }
 
     public SprintBacklog getSprintBacklogById(String id) {
-        return sprintBacklogRepository.findById(UUID.fromString(id)).orElseThrow(() -> new NotFoundException("Sprint Backlog not found"));
+        SprintBacklog sprintBacklog =  sprintBacklogRepository.findById(UUID.fromString(id))
+                .orElseThrow(() -> new NotFoundException("Sprint Backlog not found"));
+        if(!projectSecurityService.isProjectMember(sprintBacklog.getProject().getId().toString())
+        && !projectSecurityService.isProjectOwner(sprintBacklog.getProject().getId().toString())) {
+            throw new AccessDeniedException("Sprint Backlog not found");
+        }
+        return sprintBacklog;
     }
 
     public void deleteSprintBacklog(String id) {
-        getSprintBacklogById(id);
+        SprintBacklog sprintBacklog = getSprintBacklogById(id);
+        if (sprintBacklog == null) {
+            throw new NotFoundException("Sprint Backlog not found");
+        }
+        if(!projectSecurityService.isProjectOwner(sprintBacklog.getProject().getId().toString())) {
+            throw new AccessDeniedException("Sprint Backlog not found");
+        }
         sprintBacklogRepository.deleteById(UUID.fromString(id));
     }
 
     public SprintBacklog updateSprintBacklog(String id, SprintBacklogPayload payload) {
         SprintBacklog sprintBacklog = getSprintBacklogById(id);
+        if (sprintBacklog == null) {
+            throw new NotFoundException("Sprint Backlog not found");
+        }
+        if(!projectSecurityService.isProjectOwner(sprintBacklog.getProject().getId().toString())) {
+            throw new AccessDeniedException("Sprint Backlog not found");
+        }
         return sprintBacklogRepository.save(payload.toEntity(sprintBacklog));
     }
 }
